move db context, create migrations, refactor schema
