+++ Phonebook.KamilKolanowski/appsettings.json
@@ -1,0 +1,5 @@
+{
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=localhost;Database=PhoneBook;User Id=sa;Password=dockerStrongPwd123;Encrypt=False;TrustServerCertificate=True;"
+  }
+}
